third_party/WebKit/LayoutTests/fast/css/fontfaceset-events-expected.txt

CONSOLE WARNING: document.fonts.ready() method is going to be replaced with document.fonts.ready attribute in future releases. Please be prepared. For more help, check https://code.google.com/p/chromium/issues/detail?id=392077#c3 .
Test that the event handlers of FontFaceSet are called in the correct order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.type is "loading"
PASS event.type is "loadingdone"
PASS numberOfEvents is 2
PASS successfullyParsed is true

TEST COMPLETE
Hello, world!
